backend/requirements.txt
fastapi>=0.95.0
pydantic>=1.10.7
uvicorn>=0.21.1
sqlalchemy>=2.0.0
pydantic-settings==2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
aiohttp>=3.8.4
ccxt>=3.0.0
cryptography>=40.0.0
email-validator>=2.0.0
python-jose==3.3.0
requests==2.31.0
websockets>=10.4
aiosqlite==0.19.0
alembic>=1.7.7
pytest>=7.3.1
pytest-asyncio==0.21.1
httpx>=0.24.0
python-binance==1.0.19
discord.py>=2.0.0,<2.1.0
asyncpg==0.28.0
pandas==2.1.3
numpy==1.26.2
pytz==2023.3
pytest-cov==4.1.0
google-auth-oauthlib==1.2.0
google-auth==2.27.0
requests-oauthlib==1.3.1
ntplib==0.4.0
tenacity>=8.2.2
asyncio>=3.4.3
aiodns>=3.0.0
cchardet>=2.1.7
brotlipy>=0.7.0
orjson>=3.8.10
google-generativeai>=0.3.0

# WebSocket和Redis支持
aioredis==1.3.1
redis==4.5.5
hiredis==2.2.3
websockets==10.4

# 监控和度量
prometheus-client==0.16.0
prometheus-fastapi-instrumentator==6.0.0 